requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
joblib==1.0.1
kafka-python==2.0.2
kaggle==1.5.12
nltk==3.6.1
numpy==1.20.2
pandas==1.2.4
pandas-explode==0.0.6
py4j==0.10.9
pydoop==2.0.0
pyhadoop==0.1
pyspark==3.1.1
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
smart-open==5.0.0
sonopy==0.1.2
text-unidecode==1.3
textblob==0.15.3
threadpoolctl==2.1.0
tqdm==4.60.0
urllib3==1.26.4
vader==0.0.2
vaderSentiment==3.3.2
